Session Cart Implemented
Using sessions, a cart is now implemented and stores the user's input quantity for each item in an array. The array stores the product code with the quantity, and is displayed in plain text currently.

There is also validation for the input to confirm that is an integer, and will return an error message upon entering  anything other than an int.

There is also a button at the top of the cart that clears the cart array when pressed.

Test Results:
 - [Y] The product code and quantity get stored/added when input by the customer
 - [Y] The cookies can be cleared as a backup method of clearing the cart
 - [Y] There is a dedicated button clearing the cart
 - [Y] Input text is validated so that it can only be an integer