Final ReadMe Update + UpdateProductQuantity Function Update
Added comments throughout the scripts and updated the ReadMe file to contain the "How It Works" section.

In order to be more compliant with the stipulation of "A method that takes the product code(s) as a parameter and adds that to the cart", I updated the UpdateProductQuantity to take the product code as an input instead of getting the code internally.